move pipeline to projects
